move some bundled libraries no new lisp/bundled/ directory
This is done to avoid putting potentially outdated versions on the
`load-path' where they might end up shadowing more recent upstream
versions.

The new file lisp/bundled/.nosearch prevents
`normal-top-level-add-subdirs-to-load-path' from
adding the containing directory to the load path.

When requiring a FEATURE that was moved to the new lisp/bundled/
directory, we no longer use

  (require 'feature)

Instead we use

  (or (require 'feature nil t)
      (require 'feature "bundled/feature"))

This first tries to load `feature' from a library that can be found on
the `load-path' and if and only if that fails it tries to load it from a
library that can be found in a directory bundled/ that is a subdirectory
of a directory on the load path.