Fix non-advised functions not to use ems-interactive-p
ems-interactive-p is introduced to determine if a function advised by
emacspeak is called interactively; see
http://tvraman.github.io/emacspeak/blog/ems-interactive-p.html .
However in emacspeak-eterm this was used for several
emacspeak-specific functions that have no advice.  As a result,
several terminal mode commands have been silent since ems-interactive-p
was introduced.

Use an alternate mechanism for determining if a commend is called
interactively.  The documentation for called-interactively-p
recommends using an argument to the interactive special form rather
than calling called-interactively-p.  Where it was easy to do so, I
did.  However in most cases, I used called-interactively-p to avoid
bad interactions with existing interactive argument handling.