Rename nm.el to emacspeak-nm.el
A package nm exists, which also provides the feature nm.  So when
emacspeak-dbus does (require 'nm), it is undefined which of these
libraries is loaded if nm is also installed.

By renaming our nm.el to emacspeak-nm.el and requiring the respective
feature instead, we can ensure that the expected library is loaded.

That does of course not prevent something else from additionally
loading the other nm, which is problematic because both libraries
use the "nm-" symbol namespace, but at least this improves the
situation somewhat.