Re-apply the allocation removal from PR #59
This partially reverts commit f5750cca2939cefeb070ea51f1cfcf2840ba9e24.

The process of adding in the caps fix appears to have reverted some of the
changes introduced in PR #59 that got rid of the need to take a copy of
every messaged for the purpose of cloning ssml tags.

This commit restores those changes and other related reformating and
improvements to error reporting.