[espeak] Adjust espeak pitch settings to use a percentage of default
Note: the patch as currently written does meak some of the voices,
such as the higher pitch for quoted text, less distinct. This can be
adjusted going forward.

This change is a precursor to allowing the espeak voice to be
switched.

The volume settings are unaffected since volume is rarely
voice-dependent.