Add support for selecting voices to set_lang in espeak
Now, setting the language to en-us:whisper will set the language to
en-us and select the "whisper" voice, while using ":whisper" will set
the whisper voice for the current language.

Also adds a command to report the espeak data path bath to the TCL
driver, enabling the creation of a list of available voice variants to
simplify the syntax.

Finally, set up the initial language -> voice mapping using the
priority information encoded in espeak's voice list, roughly matching
the behavior of the `espeak` program.