This fixes "thrashing" audiocore on mac
After a lot of investigation, I was able
to trace the mac "distortion" was due to
rapid allocation and intiialization of
new NSSound objects, but making a player
and reusing it, the problem is solved.

Additionally, I made the textClipplayer
slightly faster by having it load the
file by reference.

Added myself to authors.