Move the play program check to play function
There is no reason to check and error on a missing play program
if you are using serve or queue and routing the playing of the
audio icons through the server.

Important Note: the error line, even when moved to this location
still breaks emacspeak when run, but this lets us avoid it by
defining

(setq emacspeak-auditory-icon-function #'emacspeak-serve-auditory-icon)

or the queue varient.