This is a complete rewrite of Swiftmac.
Way too much to list offhand, but major gist is:

- Uses swift 5.5+ for actorization, way easier to deal with
  concurrency.
- Fixed all known crashing bugs.
- Switched to AVSpeechSynthesizer (one not Deprecated)

Pitch manipluation changes will require tweaks to the way
voices are generated, but this was mostly broken in the
last release as well. Voices, pitch modulation, etc are
going to need a lot of thought for Mac and Windows long
term.